refactor: ディレクトリ構成変更
