+++ src/app/posts/[page]/opengraph-image.tsx
@@ -1,4 +1,4 @@
-import { OgpWrapper } from '@/components/Ogp/OgpWrapper'
+import { OgpWrapper } from '@/components/features'
 import { OGP_FONT } from '@/constants'
 import { OGP_IMAGE_SIZE } from '@/constants/index'
 import { fetchGoogleFonts } from '@/services/fonts'
