+++ src/components/features/Hero/index.ts
@@ -1,0 +1,1 @@
+export { Hero } from './Hero'
