+++ src/components/features/Ogp/index.ts
@@ -1,0 +1,1 @@
+export { OgpWrapper } from './OgpWrapper'
