+++ src/components/features/PostArea/index.ts
@@ -1,0 +1,1 @@
+export { PostArea } from './PostArea'
