+++ src/components/features/PostCard/index.ts
@@ -1,0 +1,1 @@
+export { PostCard } from './PostCard'
