+++ src/components/features/PostContainer/index.ts
@@ -1,0 +1,1 @@
+export { PostContainer } from './PostContainer'
