+++ src/components/ui/heading/index.ts
@@ -1,0 +1,1 @@
+export { Heading } from './Heading'
