+++ src/components/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './avatar'
+export * from './link'
+export * from './heading'
