+++ src/components/ui/link/index.ts
@@ -1,0 +1,2 @@
+export { IconLink } from './IconLink'
+export { TextLink } from './TextLink'
